requirements.txt
Django==3.2.19
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.0.0
Pillow==9.5.0
python-decouple==3.8
